+++ packages/peeky-utils/src/console.ts
@@ -1,0 +1,9 @@
+import chalk from 'chalk'
+import { isWindows } from './env.js'
+
+export function italic (text: string) {
+  if (isWindows) {
+    return text
+  }
+  return chalk.italic(text)
+}