+++ packages/peeky-utils/src/index.ts
@@ -1,4 +1,6 @@
 export * from './code.js'
+export * from './console.js'
+export * from './env.js'
 export * from './esbuild.js'
 export * from './format.js'
 export * from './fs.js'